requirements.txt
PyPDF2==2.11.1
python-docx==1.0.1
PIL==10.0.1
python-magic==0.4.27
pywin32==305.1